requirements.txt
streamlit
python-dotenv
requests
langgraph>=0.2.23
langchain>=0.2.12
langchain-openai>=0.1.17
pydantic>=2.5
typing_extensions>=4.9
google-search-results>=2.4.2
tqdm>=4.66.4
